+++ app/presenters/statistics_presenter.rb
@@ -17,10 +17,8 @@ class StatisticsPresenter
     if AppConfig.privacy.statistics.comment_counts?
       result['local_comments'] = self.local_comments
     end
-
     AppConfig.services.each do |service, options|
-      result[service] = options ? !!options["enable"] : false
-    
+      result[service] = AppConfig["services.#{service}.enable"]
     end
 
     result